vagamiga/backend/requirements.txt
Django==4.2.21
djangorestframework==3.15.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1